rework number of envelopes assigned to each atom
An exponential envelope is assigned to each atom for every occupied
shell it has. It is assumed that the main shells are occupied in strict
increasing order (not true for e.g. transition metals). For example
carbon has an occupation of 1s2 2(s2 p2), therefore it receives two
envelopes, their zeta coefficients initialized to Z/1 and Z/2.