refactor edge building
The edges are now built with the common EdgeFactory class defined in
graph.py. Upon initialization it expects a list of labels, denoting the
type of edges that need to be constructed. The type of edges required by
the different message passing architectures (SchNet, PaiNN) can be
accessed through the required_edge_types class method on these classes.