use hk.Embed for initial electronic embeddings
The default KFAC dense curvature approximation yields NaNs when
using a linear layer for the initial embeddings of the electrons,
due to the fact that these embeddings are the same for each sample
in a batch. When using hk.Embed, KFAC employs a trivial diagonal
curvature approximation, which avoids NaNs.