implement general wrapper for overflow checking
The function deepqmc.jax.utils.check_overflow can be used to wrap a
function. During exectution it checks whether the output of the wrapped
function indicates an overflow, and if so it reruns the function until
there is no overflow. The wrapped function has to satisfy the expected
signature:
smpl_state, *other = func(rng, smpl_state, *arts, **kwargs)