represent molecules with a frozen dataclass
In some parts of the Ansatz, we might need additional data about
the molecule, like the number of spin-up electrons, that can be
calculated from the specification of the molecule (charges, charge,
spin). However, this calculation cannot be done inside jitted regions,
due to limitations of jax.jit when working with bare ints or floats.
Therefore we need to precalculate these quantities and store them
in the Molecule object. To avoid the usual problems associated with
these constructs, we need the Molecule object to be immutable.