remove DenseBlock from kfacext
This block cuvature estimator was used to handle layers that didn't have
a first dimension of size `batch_dim`. This is usually the result of a
layer that acts on inputs that are independent from the sample index,
e.g. the initial electron embeddings are the same for all samples.
From now on, these types of layers should be represented with an
hk.Embed layer, the curvature of which is calculated with a diagonal
approximation.