make wf_state argument of sampler.init optional
Since initializing a sampler with no available preexisting WF state
has become a common pattern, it makes sense to handle this case within
the `init` function of the samplers. Now, the `wf_state` argument of the
`sampler.init` functions can be omitted, in which case a new valid WF
state is created as part of the sampler initialization.