match positional_electron_embedding to FermiNet
If positional_electron_embedding is `True`, don't use the
ElectronEmbeddings embedding layer, but use the nuclei-electron edge
features directly as initial electron embeddings. The computation of
these edge features is decoupled from the main edge features of the
graph, to make it clear that these are used for node embeddings and not
as usual edges.