make w for ne edges optional
This is by far the hackiest changes for DeepErwin. Contrary to their
manuscript they (unintentionally?) don't use a `w` network (`B` network
in the notation of their manuscript) for the nuclei-electron edges. We
can now mimic this by setting the `layer_factory` config option
`w_for_ne` to `False`. This is such a weird choice that I've kept the
default `w_for_ne=True` in the code, so that we don't have to specify it
in all other config files.