add softplus_zeta option
Instead of taking the absolute value of the envelope exponents,
DeepErwin passes the `zeta` parameter through `softplus`. This means
that by default KFAC can't recognize this layer as shift_and_scale. The
manual registration of this layer is not implemented for us, e.g. if
`softplus_zeta=True`, the zeta parameter will be `Orphan` with KFAC.