fix chained restarts
When a run is restarted multiple times, the original config file can be
found in the workdir of the very first run. In such a case we follow the
runs backwards, looking in the `restdir` of each of them, until we find
a run that doesn't have a `restdir`, which is assumed to be the first
run we are looking for.